+++ docs/history.md
@@ -50,6 +50,39 @@
 
 --------------------------------------------------------------------------------
 
+[![P]][H] **v0.0.6 (dev)**
+--------------------------------------------------------------------------------
+[![V]][VE006] [![2025-11-11 23:50]][VE006]  
+
+[2025-11-11 23:50]: https://img.shields.io/static/v1?label=2025-11-11&message=23:50&color=yellowgreen
+[VE006]: changelog.md#-v006-dev
+
+|    дата    | время |       ветка      | статус |  
+|:----------:|:-----:|:----------------:|:------:|  
+| 2025-11-11 | 23:40 | [#6-dev-support] | DONE   |  
+| 2025-11-11 | 04:30 | [#6-dev-support] | BEGIN  |  
+
+1. upd: `docs`  
+   - upd: `logo.png`     (006) PRE  
+   - upd: `changelog.md` (006) PRE  
+   - upd: `chrono.md`    (006) PRE  
+   - upd: `history.md`   (006) PRE  
+   - upd: `docs.md`      (006) PRE  
+2. add: `docs/docs`  
+   - add: `wchar.md`     (001) PRE  
+   - add: `cstdint.md`   (001) PRE  
+   - add: `support.md`   (009) PRE  
+3. upd: `include/compile`  
+   - upd: `compile.ver`  (005) PRE  
+   - add: `support.hpp`  (009) PRE  
+   - add: `cstdint.hpp`  (002) PRE  
+   - add: `wchar.hpp`    (002) PRE  
+4. upd: `project.root`   (006) PRE  
+
+[#6-dev-support]: tasks/2025-11-11-0006-dev-support.md
+<div/>
+
+
 [![S]][H] **v0.0.5 (dev)**
 --------------------------------------------------------------------------------
 [![V]][VE005] [![2025-11-11 04:30]][VE005]